backupccl: exclude dropped descriptors from user-descriptor check
During a full cluster restore, we check for user created descriptors.
Previously, this check would fail on any user-defined descriptor
because we restored most descriptors as-is without any rewriting.
However, we now rewrite all descriptors during restore.

Because of this, we can limit our check to descriptors that are not
dropped and thus might produce namespace conflicts with descriptors we
are going to restore.

This allows us to more quickly retry a failed or cancelled backup
since the user does not have to wait for descriptor to be completely
deleted by the GC job.

Release note: None